fix: corrected and refactored cms dataflow in (profile)
