+++ package.json
@@ -5,6 +5,8 @@
   "scripts": {
     "dev": "next dev",
     "build": "next build",
+    "serve": "npx serve@latest out",
+    "build:serve": "next build && npx serve@latest out",
     "postbuild": "next-sitemap && node buildScripts/postbuild.js",
     "start": "next start",
     "lint": "next lint --fix"
